test: add folders for Unit/Integration tests
- Fix the namespace of test classes.
