+++ tests/AppBundle/Unit/Controller/UserControllerTest.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Tests\AppBundle\Controller;
+namespace Tests\AppBundle\Unit\Controller;
 
-use AppBundle\Controller\TaskController;
-use AppBundle\Entity\Task;
+use AppBundle\Controller\UserController;
 use AppBundle\Entity\User;
-use AppBundle\Form\TaskType;
+use AppBundle\Form\UserType;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Framework\TestCase;
@@ -16,10 +15,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
 use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
-class TaskControllerTest extends TestCase
+class UserControllerTest extends TestCase
 {
     /**
      * This test checks that the method listAction() is correctly returned
@@ -34,51 +32,21 @@ class TaskControllerTest extends TestCase
     {
         $objectRepository = $this->createMock(ObjectRepository::class);
         $objectRepository->expects($this->once())
-            ->method('findBy');
+            ->method('findAll');
 
         $entityManager = $this->createMock(EntityManagerInterface::class);
         $entityManager->expects($this->once())
             ->method('getRepository')
-            ->with('AppBundle:Task')
+            ->with('AppBundle:User')
             ->willReturn($objectRepository);
 
-        $twig = $this->getTwigMock(true, 'task/list.html.twig', [
-            'tasks' => null,
+        $twig = $this->getTwigMock(true, 'user/list.html.twig', [
+            'users' => null,
         ]);
 
-        $controller = new TaskController($twig, $entityManager);
+        $controller = new UserController($twig, $entityManager);
 
         $this->assertInstanceOf(Response::class, $controller->listAction());
-    }
-
-    /**
-     * This test checks that the method listDoneAction() is correctly returned
-     * and checks that the methods are correctly called.
-     *
-     * @throws \ReflectionException
-     * @throws \Twig_Error_Loader
-     * @throws \Twig_Error_Runtime
-     * @throws \Twig_Error_Syntax
-     */
-    public function testListDone()
-    {
-        $objectRepository = $this->createMock(ObjectRepository::class);
-        $objectRepository->expects($this->once())
-            ->method('findBy');
-
-        $entityManager = $this->createMock(EntityManagerInterface::class);
-        $entityManager->expects($this->once())
-            ->method('getRepository')
-            ->with('AppBundle:Task')
-            ->willReturn($objectRepository);
-
-        $twig = $this->getTwigMock(true, 'task/list_done.html.twig', [
-            'tasks' => null,
-        ]);
-
-        $controller = new TaskController($twig, $entityManager);
-
-        $this->assertInstanceOf(Response::class, $controller->listDoneAction());
     }
 
     /**
@@ -96,41 +64,33 @@ class TaskControllerTest extends TestCase
      */
     public function testCreate(bool $formIsSubmitted, string $className)
     {
-        $task = new Task();
-
-        $token = $this->createMock(TokenInterface::class);
-        $token->expects($this->once())
-            ->method('getUser')
-            ->willReturn(new User());
-
-        $tokenStorage = $this->createMock(TokenStorageInterface::class);
-        $tokenStorage->expects($this->once())
-            ->method('getToken')
-            ->willReturn($token);
+        $user = new User();
 
         $request = $this->createMock(Request::class);
 
         $form = $this->getFormMock($formIsSubmitted, $request);
 
-        $formFactory = $this->getFormInterfaceMock($form);
+        $formFactory = $this->getFormInterfaceMock($form, $user);
 
-        $entityManager = $this->getEntityManagerMock($formIsSubmitted, false, $formIsSubmitted, $task);
+        $passwordEncoder = $this->getUserPasswordEncoderMock($formIsSubmitted, $user);
 
-        $flashBag = $this->getFlashBagMock($formIsSubmitted, 'success', 'La tâche a bien été ajoutée.');
+        $entityManager = $this->getEntityManagerMock($formIsSubmitted, false, $formIsSubmitted, $user);
 
-        $router = $this->getRouterMock($formIsSubmitted, 'task_list');
+        $flashBag = $this->getFlashBagMock($formIsSubmitted, 'success', 'L\'utilisateur a bien été ajouté.');
 
-        $twig = $this->getTwigMock(!$formIsSubmitted, 'task/create.html.twig', [
+        $router = $this->getRouterMock($formIsSubmitted, 'user_list');
+
+        $twig = $this->getTwigMock(!$formIsSubmitted, 'user/create.html.twig', [
             'form' => null,
         ]);
 
-        $controller = new TaskController($twig, $entityManager);
+        $controller = new UserController($twig, $entityManager);
 
-        $this->assertInstanceOf($className, $controller->createAction($tokenStorage, $formFactory, $request, $flashBag, $router));
+        $this->assertInstanceOf($className, $controller->createAction($formFactory, $request, $passwordEncoder, $flashBag, $router));
     }
 
     /**
-     * This test checks that the method editAction() is correctly returned
+     * This test checks that the method EditAction() is correctly returned
      * and checks that the methods are correctly called or not.
      *
      * @param string $className
@@ -144,28 +104,30 @@ class TaskControllerTest extends TestCase
      */
     public function testEdit(bool $formIsSubmitted, string $className)
     {
-        $task = new Task();
+        $user = new User();
 
         $request = $this->createMock(Request::class);
 
         $form = $this->getFormMock($formIsSubmitted, $request);
 
-        $formFactory = $this->getFormInterfaceMock($form);
+        $formFactory = $this->getFormInterfaceMock($form, $user);
 
-        $entityManager = $this->getEntityManagerMock(false, false, $formIsSubmitted, $task);
+        $passwordEncoder = $this->getUserPasswordEncoderMock($formIsSubmitted, $user);
 
-        $flashBag = $this->getFlashBagMock($formIsSubmitted, 'success', 'La tâche a bien été modifiée.');
+        $entityManager = $this->getEntityManagerMock(false, false, $formIsSubmitted, $user);
 
-        $router = $this->getRouterMock($formIsSubmitted, 'task_list');
+        $flashBag = $this->getFlashBagMock($formIsSubmitted, 'success', 'L\'utilisateur a bien été modifié');
 
-        $twig = $this->getTwigMock(!$formIsSubmitted, 'task/edit.html.twig', [
+        $router = $this->getRouterMock($formIsSubmitted, 'user_list');
+
+        $twig = $this->getTwigMock(!$formIsSubmitted, 'user/edit.html.twig', [
             'form' => null,
-            'task' => $task,
+            'user' => $user,
         ]);
 
-        $controller = new TaskController($twig, $entityManager);
+        $controller = new UserController($twig, $entityManager);
 
-        $this->assertInstanceOf($className, $controller->editAction($formFactory, $task, $request, $flashBag, $router));
+        $this->assertInstanceOf($className, $controller->editAction($formFactory, $user, $request, $passwordEncoder, $flashBag, $router));
     }
 
     /**
@@ -177,70 +139,6 @@ class TaskControllerTest extends TestCase
             [true, RedirectResponse::class],
             [false, Response::class],
         ];
-    }
-
-    /**
-     * This test checks that the method toggleAction() is correctly returned
-     * and checks that the methods are correctly called.
-     *
-     * @param bool   $taskIsDone
-     * @param string $message
-     * @param string $route
-     *
-     * @throws \ReflectionException
-     * @dataProvider provideTaskIsDone
-     */
-    public function testToggleTask(bool $taskIsDone, string $message, string $route)
-    {
-        $task = $this->createMock(Task::class);
-
-        $task->method('isDone')->willReturn($taskIsDone);
-
-        $entityManager = $this->getEntityManagerMock(false, false, true, $task);
-
-        $flashBag = $this->getFlashBagMock(true, 'success', sprintf($message, ''));
-
-        $router = $this->getRouterMock(true, $route);
-
-        $twig = $this->getTwigMock(false);
-
-        $controller = new TaskController($twig, $entityManager);
-
-        $this->assertInstanceOf(RedirectResponse::class, $controller->toggleTaskAction($task, $flashBag, $router));
-    }
-
-    /**
-     * @return array
-     */
-    public function provideTaskIsDone()
-    {
-        return [
-            [true, 'La tâche %s a bien été marquée comme faite.', 'task_list'],
-            [false, 'La tâche %s a bien été marquée comme non terminée.', 'task_done'],
-        ];
-    }
-
-    /**
-     * This test checks that the method deleteAction() is correctly returned
-     * and checks that the methods are correctly called.
-     *
-     * @throws \ReflectionException
-     */
-    public function testDeleteTask()
-    {
-        $task = new Task();
-
-        $entityManager = $this->getEntityManagerMock(false, true, true, $task);
-
-        $flashBag = $this->getFlashBagMock(true, 'success', 'La tâche a bien été supprimée.');
-
-        $router = $this->getRouterMock(true, 'task_list');
-
-        $twig = $this->getTwigMock(false);
-
-        $controller = new TaskController($twig, $entityManager);
-
-        $this->assertInstanceOf(RedirectResponse::class, $controller->deleteTaskAction($task, $flashBag, $router));
     }
 
     /**
@@ -270,17 +168,18 @@ class TaskControllerTest extends TestCase
      * and checks that create() is correctly called.
      *
      * @param $form
+     * @param User $user
      *
      * @throws \ReflectionException
      *
      * @return \PHPUnit\Framework\MockObject\MockObject|FormFactoryInterface
      */
-    private function getFormInterfaceMock($form)
+    private function getFormInterfaceMock($form, User $user)
     {
         $formFactoryMock = $this->createMock(FormFactoryInterface::class);
         $formFactoryMock->expects($this->once())
             ->method('create')
-            ->with(TaskType::class)
+            ->with(UserType::class, $user)
             ->willReturn($form);
 
         return $formFactoryMock;
@@ -314,26 +213,48 @@ class TaskControllerTest extends TestCase
     }
 
     /**
+     * This helper method mocks UserPasswordEncoderInterface
+     * and checks that encodePassword() is correctly called or not called.
+     *
+     * @param bool $callEncodePassword
+     * @param User $user
+     *
+     * @throws \ReflectionException
+     *
+     * @return \PHPUnit\Framework\MockObject\MockObject|UserPasswordEncoderInterface
+     */
+    private function getUserPasswordEncoderMock(bool $callEncodePassword, User $user)
+    {
+        $userPasswordEncoder = $this->createMock(UserPasswordEncoderInterface::class);
+        $userPasswordEncoder->expects($callEncodePassword ? $this->once() : $this->never())
+            ->method('encodePassword')
+            ->with($user, $user->getPassword());
+
+        return $userPasswordEncoder;
+    }
+
+    /**
      * This helper method mocks EntityManagerInterface
      * and checks that persist(), remove(), flush() are correctly called or not called.
      *
      * @param bool $callPersist
      * @param bool $callRemove
      * @param bool $callFlush
-     * @param Task $task
+     * @param User $user
      *
      * @throws \ReflectionException
      *
      * @return \PHPUnit\Framework\MockObject\MockObject|EntityManagerInterface
      */
-    private function getEntityManagerMock(bool $callPersist, bool $callRemove, bool $callFlush, Task $task)
+    private function getEntityManagerMock(bool $callPersist, bool $callRemove, bool $callFlush, User $user)
     {
         $entityManagerMock = $this->createMock(EntityManagerInterface::class);
         $entityManagerMock->expects($callPersist ? $this->once() : $this->never())
-            ->method('persist');
+            ->method('persist')
+            ->with($user);
         $entityManagerMock->expects($callRemove ? $this->once() : $this->never())
             ->method('remove')
-            ->with($task);
+            ->with($user);
         $entityManagerMock->expects($callFlush ? $this->once() : $this->never())
             ->method('flush');
